requirements.txt
Django~=4.1.7
django-allauth
django-bootstrap5
django-bootstrap-icons
django-extensions
django-pwa
paho-mqtt

